+++ CreateGame/src/test/java/org/gamehunter/creategame/objects/SimpleCardTests.java
@@ -6,7 +6,7 @@ import org.gamehunter.creategame.constants.GameAreaName;
 import org.gamehunter.creategame.locations.GameLocation;
 import org.gamehunter.creategame.objects.characteristics.ConcreteCharacteristic;
 import org.gamehunter.creategame.values.ArbitraryValue;
-import org.gamehunter.creategame.values.IntegerValue;
+import org.gamehunter.creategame.values.DoubleValue;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -28,7 +28,7 @@ public class SimpleCardTests {
 
     SimpleCard createOneOfClubs() {
         SimpleCard oneOfClubs = new SimpleCard(new GameLocation(GameAreaName.COMMON));
-        oneOfClubs.addCharacteristic(new ConcreteCharacteristic("Value", new IntegerValue(1)))
+        oneOfClubs.addCharacteristic(new ConcreteCharacteristic("Value", new DoubleValue(1)))
         .addCharacteristic(new ConcreteCharacteristic("Suit", new ArbitraryValue<String>("Clubs")));
         return oneOfClubs;
     }
